+++ packages/dropdown/src/Dropdown.tsx
@@ -1,12 +1,18 @@
 /* eslint-disable prettier/prettier */
-import React, { useState } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import cn from 'classnames';
 
-import { DropdownProps, SelectOptions } from './Dropdown.types';
+import { ChevronDown, ChevronTop } from '@heycar-uikit/icons';
+import Input from '@heycar-uikit/input';
+
+import DropdownOption from './components/DropdownOption';
+import { DropdownOptionProps } from './components/DropdownOption.types';
+import { DropdownProps } from './Dropdown.types';
 
 import styles from './styles/default.module.css';
 
-export const Dropdown = ({
+function Dropdown({
+  Component = 'div',
   value,
   onChange,
   options,
@@ -14,22 +20,72 @@ export const Dropdown = ({
   dataTestId,
   onBlur,
   onClick,
+  label,
+  hint,
   fullWidth,
-}: DropdownProps) => {
-  const [stateValue, setStateValue] = useState<SelectOptions | undefined>(
+  placeholder,
+  error,
+  ...restProps
+}: DropdownProps) {
+  const [stateValue, setStateValue] = useState<DropdownOptionProps | undefined>(
     value,
   );
 
   const [isOpen, setIsOpen] = useState(false);
+  const [highlightedIndex, setHighlightedIndex] = useState(0);
+  const containerRef = useRef<HTMLDivElement>(null);
 
-  const selectOption = (option: SelectOptions) => {
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  const selectOption = (option: DropdownOptionProps) => {
     if (onChange) {
       if (option !== stateValue) onChange(option);
     }
     setStateValue(option);
   };
 
-  const isOptionSelection = (option: SelectOptions) => {
+  useEffect(() => {
+    if (isOpen) setHighlightedIndex(0);
+  }, [isOpen]);
+
+  useEffect(() => {
+    const handler = (e: KeyboardEvent) => {
+      if (e.target != containerRef.current) return;
+      switch (e.code) {
+        case 'Enter':
+        case 'Space':
+          setIsOpen(prev => !prev);
+          if (isOpen) selectOption(options[highlightedIndex]);
+          break;
+        case 'ArrowUp':
+        case 'ArrowDown': {
+          if (!isOpen) {
+            setIsOpen(true);
+            break;
+          }
+
+          const newValue = highlightedIndex + (e.code === 'ArrowDown' ? 1 : -1);
+
+          if (newValue >= 0 && newValue < options.length) {
+            setHighlightedIndex(newValue);
+          }
+          break;
+        }
+        case 'Escape':
+          setIsOpen(false);
+          break;
+      }
+    };
+
+    containerRef.current?.addEventListener('keydown', handler);
+
+    return () => {
+      // eslint-disable-next-line react-hooks/exhaustive-deps
+      containerRef.current?.removeEventListener('keydown', handler);
+    };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [highlightedIndex, options, selectOption]);
+
+  const isOptionSelection = (option: DropdownOptionProps) => {
     return option?.value === stateValue?.value;
   };
 
@@ -48,59 +104,53 @@ export const Dropdown = ({
     fullWidth && styles.fullWidth,
   );
 
-  const arrowClassNames = cn(
-    disabled && styles.caret_disabled,
-    isOpen && styles.caret_up,
-    !isOpen && styles.caret_down,
-  );
-
-  const valueClassNames = cn(
-    disabled && 'disabled',
-    styles.value,
-  );
+  const valueClassNames = cn(disabled && 'disabled', styles.value);
 
   return (
-    <div
+    <Component
       className={classNames}
-      onBlur={onBlurHandler}
       onClick={onClickHandler}
+      ref={containerRef}
       tabIndex={0}
+      {...restProps}
+      onBlur={onBlurHandler}
     >
-      <span className={valueClassNames}>
-        {stateValue?.label}
-      </span>
-      <div className={arrowClassNames}></div>
+      <Input
+        className={valueClassNames}
+        disabled={disabled}
+        error={error}
+        fullWidth={true}
+        hint={hint}
+        label={label}
+        onChange={() => {
+          setStateValue(options[highlightedIndex]);
+        }}
+        placeholder={placeholder}
+        rightIcon={isOpen ? <ChevronTop /> : <ChevronDown />}
+        value={stateValue?.label}
+      />
       <ul
         className={`${styles.options} ${isOpen ? styles.show : ''}`}
         data-test-id={dataTestId}
       >
-        {options.map(option => (
+        {options.map((option, index) => (
           <li
             className={`${styles.option} ${isOptionSelection(option) ? styles.selected : ''
-              }`}
+              } ${index === highlightedIndex ? styles.highlighted : ''}`}
             key={option.value}
-            onClick={e => {
-              e.stopPropagation();
+            onMouseDown={e => {
               selectOption(option);
+              e.stopPropagation();
               setIsOpen(false);
             }}
+            onMouseEnter={() => setHighlightedIndex(index)}
           >
-            <div className={styles.contentContainer}>
-              {option.leftContent && (
-                <span className={styles.leftContent}>{option.leftContent}</span>
-              )}
-              <span className={styles.optionContent}>{option.label}</span>
-              {option.rightContent && (
-                <span className={styles.rightContent}>
-                  {option.rightContent}
-                </span>
-              )}
-            </div>
+            <DropdownOption {...option} />
           </li>
         ))}
       </ul>
-    </div>
+    </Component>
   );
-};
+}
 
 export default Dropdown;