+++ packages/dropdown/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.3](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/dropdown@2.0.2...@heycar-uikit/dropdown@2.0.3) (2023-02-16)
+
+
+### Bug Fixes
+
+* **dropdown:** adds error state ([#140](https://github.com/hey-car/heycar-uikit/issues/140)) ([4ef925f](https://github.com/hey-car/heycar-uikit/commit/4ef925f3cc4db119a3dc3b0da7c7566d5ba4bcac))
+
+
+
+
+
+## [2.0.2](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/dropdown@2.0.1...@heycar-uikit/dropdown@2.0.2) (2023-02-15)
+
+
+### Bug Fixes
+
+* **dropdown:** adding accessibility ([#137](https://github.com/hey-car/heycar-uikit/issues/137)) ([5e8cbec](https://github.com/hey-car/heycar-uikit/commit/5e8cbec6755771216244e36675f8ff7c79bcbfb5))
+
+
+
+
+
+## [2.0.1](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/dropdown@2.0.0...@heycar-uikit/dropdown@2.0.1) (2023-01-30)
+
+
+### Bug Fixes
+
+* **dropdown:** add rest props update styles ([#132](https://github.com/hey-car/heycar-uikit/issues/132)) ([e940396](https://github.com/hey-car/heycar-uikit/commit/e94039645dee00dd18711fe4256636dea49969d9))
+
+
+
+
+
 # [2.0.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/dropdown@1.0.2...@heycar-uikit/dropdown@2.0.0) (2023-01-17)